requirements.txt
customtkinter
google-generativeai
google-cloud-texttospeech
pillow
